requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg[binary]==3.3.2
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.9
email-validator==2.1.0
pydantic==2.12.5
python-dotenv==1.0.1